CI: Use actions/setup-go@v4
The action will now read the major and minor version from the go.mod file and will use the latest patch of that version.
Also, caching of modules and build outputs is now enabled by default. It uses actions/cache under the hood, but simplifies the configuration, e.g. by automatically looking up the go.sum file in the projects root directory and using its hash as part of the cache-key.